requirements.txt
psutil==5.9.8
psycopg2-binary==2.9.9
python-dotenv==1.0.1
rich==13.7.0
SQLAlchemy==2.0.27